second commit

+++ src/tssclubfair2017problemb/TSSClubFair2017ProblemB.java
@@ -16,7 +16,7 @@ public class TSSClubFair2017ProblemB {
     /**
      * @param args the command line arguments
      */
-    static double[] distances;
+    static long[] distances;
     public static void main(String[] args) {
         // TODO code application logic here
         Scanner sc=new Scanner(System.in);
@@ -25,17 +25,17 @@ public class TSSClubFair2017ProblemB {
         int n=Integer.parseInt(pair[0]);
         int q=Integer.parseInt(pair[1]);
 
-        distances=new double[n];
+        distances=new long[n];
         for (int i=0;i<n;i++) {
             line=sc.nextLine();
             pair=line.split(" ");
-            double x=(double)Integer.parseInt(pair[0]);
-            double y=(double)Integer.parseInt(pair[1]);
+            long x=Long.parseLong(pair[0]);
+            long y=Long.parseLong(pair[1]);
             distances[i]=x*x+y*y;
         }
         for (int i=0;i<q;i++) {
             line=sc.nextLine();
-            double impactDistance=(double)Integer.parseInt(line);
+            long impactDistance=Long.parseLong(line);
             impactDistance*=impactDistance;
             int count=0;
             for (int j=0;j<n;j++) {
